[refactor]: 최근 검색어 로직 파일 이동 (#83)
