+++ app/src/main/java/com/texthip/thip/data/repository/BookRepository.kt
@@ -7,7 +7,6 @@ import com.texthip.thip.data.model.book.response.BookListResponse
 import com.texthip.thip.data.model.book.response.BookSaveResponse
 import com.texthip.thip.data.model.book.response.BookSearchResponse
 import com.texthip.thip.data.model.book.response.MostSearchedBooksResponse
-import com.texthip.thip.data.model.book.response.RecentSearchResponse
 import com.texthip.thip.data.model.book.response.RecruitingRoomsResponse
 import com.texthip.thip.data.service.BookService
 import javax.inject.Inject
@@ -35,20 +34,6 @@ class BookRepository @Inject constructor(
     /** 인기 책 조회 */
     suspend fun getMostSearchedBooks(): Result<MostSearchedBooksResponse?> = runCatching {
         bookService.getMostSearchedBooks()
-            .handleBaseResponse()
-            .getOrThrow()
-    }
-
-    /** 최근 검색어 조회 */
-    suspend fun getRecentSearches(type: String = "BOOK"): Result<RecentSearchResponse?> = runCatching {
-        bookService.getRecentSearches(type)
-            .handleBaseResponse()
-            .getOrThrow()
-    }
-
-    /** 최근 검색어 삭제 */
-    suspend fun deleteRecentSearch(recentSearchId: Int): Result<Unit> = runCatching {
-        bookService.deleteRecentSearch(recentSearchId)
             .handleBaseResponse()
             .getOrThrow()
     }
